feat: Remove bi.webview* flag in code
Since now the bi webviews are fully deployed

Also removed the budget-insight policy, directly related to the previous
way of handling powens konnectors.
