+++ packages/cozy-harvest-lib/src/components/DisconnectedAccountModal.spec.jsx
@@ -1,6 +1,7 @@
 import { render } from '@testing-library/react'
 import React from 'react'
 
+import { CozyConfirmDialogProvider } from './CozyConfirmDialogProvider'
 import DisconnectedAccountModal from './DisconnectedAccountModal'
 import bankAccounts from './KonnectorConfiguration/ConfigurationTab/bank-account-fixture.json'
 import AppLike from '../../test/AppLike'
@@ -17,10 +18,12 @@ describe('DisconnectedAccountModal', () => {
     mockClient.dispatch = jest.fn()
     const root = render(
       <AppLike client={mockClient}>
-        <DisconnectedAccountModal
-          accounts={[bankAccounts]}
-          onClose={jest.fn()}
-        />
+        <CozyConfirmDialogProvider>
+          <DisconnectedAccountModal
+            accounts={[bankAccounts]}
+            onClose={jest.fn()}
+          />
+        </CozyConfirmDialogProvider>
       </AppLike>
     )
     return { root }
